Fix zip nxtheme support, layout fixes, migrate menu entry items from headers to cpp
