+++ SwitchThemesNX/SwitchThemesNX.vcxproj
@@ -73,9 +73,7 @@
     <ClInclude Include="source\Pages\RemoteInstall\Worker.hpp" />
     <ClInclude Include="source\Pages\SettingsPage.hpp" />
     <ClInclude Include="source\Pages\TextPage.hpp" />
-    <ClInclude Include="source\Pages\ThemeEntry\ImagePreview.hpp" />
     <ClInclude Include="source\Pages\ThemeEntry\ThemeEntry.hpp" />
-    <ClInclude Include="source\Pages\ThemeEntry\NxEntry.hpp" />
     <ClInclude Include="source\Pages\ThemePage.hpp" />
     <ClInclude Include="source\Pages\UninstallPage.hpp" />
     <ClInclude Include="source\Platform\Platform.hpp" />
@@ -439,8 +437,10 @@
     <ClCompile Include="source\Pages\SettingsPage.cpp" />
     <ClCompile Include="source\Pages\TextPage.cpp" />
     <ClCompile Include="source\Pages\ThemeEntry\BaseEntry.cpp" />
-    <ClCompile Include="source\Pages\ThemeEntry\FontEntry.hpp" />
-    <ClCompile Include="source\Pages\ThemeEntry\LegacyEntry.hpp" />
+    <ClCompile Include="source\Pages\ThemeEntry\FontEntry.cpp" />
+    <ClCompile Include="source\Pages\ThemeEntry\ImagePreview.hpp" />
+    <ClCompile Include="source\Pages\ThemeEntry\LegacyEntry.cpp" />
+    <ClCompile Include="source\Pages\ThemeEntry\NxEntry.cpp" />
     <ClCompile Include="source\Pages\ThemePage.cpp" />
     <ClCompile Include="source\Pages\UninstallPage.cpp" />
     <ClCompile Include="source\Platform\PlatformSwitch.cpp" />
